Updated registration details form to use ES6 class, and renamed for consistency
